+++ src/matrices/mx_transformations.c
@@ -3,18 +3,18 @@
 /*                                                        :::      ::::::::   */
 /*   mx_transformations.c                               :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yde-goes <yde-goes@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: mdias-ma <mdias-ma@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/27 10:44:35 by yde-goes          #+#    #+#             */
-/*   Updated: 2023/03/27 11:23:19 by yde-goes         ###   ########.fr       */
+/*   Updated: 2023/05/24 11:13:52 by mdias-ma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "matrices.h"
 
-t_matrix	translation(float x, float y, float z)
+t_matrix	translation(double x, double y, double z)
 {
-	const float	transl[MAX][MAX] = {
+	const double	transl[MAX][MAX] = {
 	{1, 0, 0, x},
 	{0, 1, 0, y},
 	{0, 0, 1, z},
@@ -24,9 +24,9 @@ t_matrix	translation(float x, float y, float z)
 	return (create_matrix(transl, MAX));
 }
 
-t_matrix	scaling(float x, float y, float z)
+t_matrix	scaling(double x, double y, double z)
 {
-	const float	scale[MAX][MAX] = {
+	const double	scale[MAX][MAX] = {
 	{x, 0, 0, 0},
 	{0, y, 0, 0},
 	{0, 0, z, 0},
@@ -38,7 +38,7 @@ t_matrix	scaling(float x, float y, float z)
 
 t_matrix	shearing(t_shearing x, t_shearing y, t_shearing z)
 {
-	const float	shear[MAX][MAX] = {
+	const double	shear[MAX][MAX] = {
 	{1, x.p1, x.p2, 0},
 	{y.p1, 1, y.p2, 0},
 	{z.p1, z.p2, 1, 0},
