+++ src/tuples/basic_math.c
@@ -6,7 +6,7 @@
 /*   By: mdias-ma <mdias-ma@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/17 13:25:26 by mdias-ma          #+#    #+#             */
-/*   Updated: 2023/03/17 13:25:50 by mdias-ma         ###   ########.fr       */
+/*   Updated: 2023/05/24 11:14:52 by mdias-ma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -42,7 +42,7 @@ t_tuple	negate(t_tuple a)
 	});
 }
 
-t_tuple	multiply(t_tuple a, float scalar)
+t_tuple	multiply(t_tuple a, double scalar)
 {
 	return ((t_tuple){
 		a.x * scalar,
@@ -52,7 +52,7 @@ t_tuple	multiply(t_tuple a, float scalar)
 	});
 }
 
-t_tuple	divide(t_tuple a, float scalar)
+t_tuple	divide(t_tuple a, double scalar)
 {
 	return ((t_tuple){
 		a.x / scalar,
